requirements.txt
Werkzeug==2.0.2
flask
pypdf
python-docx
pypandoc